+++ src/main/java/woowacourse/shoppingcart/domain/Address.java
@@ -1,4 +1,4 @@
-package woowacourse.auth.domain;
+package woowacourse.shoppingcart.domain;
 
 public class Address {
