+++ src/main/java/woowacourse/shoppingcart/exception/InvalidEmailException.java
@@ -1,4 +1,4 @@
-package woowacourse.auth.exception;
+package woowacourse.shoppingcart.exception;
 
 public class InvalidEmailException extends InvalidPropertyException {
