+++ src/main/java/woowacourse/shoppingcart/exception/InvalidPasswordException.java
@@ -1,4 +1,4 @@
-package woowacourse.auth.exception;
+package woowacourse.shoppingcart.exception;
 
 public class InvalidPasswordException extends InvalidPropertyException {
