+++ src/main/java/woowacourse/shoppingcart/exception/InvalidPhoneNumberException.java
@@ -1,4 +1,4 @@
-package woowacourse.auth.exception;
+package woowacourse.shoppingcart.exception;
 
 public class InvalidPhoneNumberException extends InvalidPropertyException {
