+++ src/main/java/woowacourse/shoppingcart/exception/ShoppingCartException.java
@@ -1,4 +1,4 @@
-package woowacourse.auth.exception;
+package woowacourse.shoppingcart.exception;
 
 public class ShoppingCartException extends RuntimeException {
