+++ src/test/java/woowacourse/shoppingcart/domain/AddressTest.java
@@ -1,4 +1,4 @@
-package woowacourse.auth.domain;
+package woowacourse.shoppingcart.domain;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
