+++ src/test/java/woowacourse/shoppingcart/domain/PhoneNumberTest.java
@@ -1,4 +1,4 @@
-package woowacourse.auth.domain;
+package woowacourse.shoppingcart.domain;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
@@ -9,7 +9,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import woowacourse.auth.exception.InvalidPhoneNumberException;
+import woowacourse.shoppingcart.exception.InvalidPhoneNumberException;
 
 class PhoneNumberTest {
